requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements.txt requirements.in
appnope==0.1.4
    # via ipykernel
asttokens==3.0.0
    # via stack-data
comm==0.2.3
    # via ipykernel
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.15
    # via ipykernel
decorator==5.2.1
    # via ipython
executing==2.2.0
    # via stack-data
faker==37.4.2
    # via -r requirements.in
fonttools==4.59.0
    # via matplotlib
ipykernel==6.30.0
    # via -r requirements.in
ipython==9.4.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
kiwisolver==1.4.8
    # via matplotlib
matplotlib==3.10.3
    # via -r requirements.in
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
neo4j==5.28.1
    # via -r requirements.in
nest-asyncio==1.6.0
    # via ipykernel
networkx==3.5
    # via -r requirements.in
numpy==2.3.2
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
packaging==25.0
    # via
    #   ipykernel
    #   matplotlib
pandas==2.3.1
    # via -r requirements.in
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.3.0
    # via matplotlib
platformdirs==4.3.8
    # via jupyter-core
prompt-toolkit==3.0.51
    # via ipython
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via
    #   neo4j
    #   pandas
pyzmq==27.0.0
    # via
    #   ipykernel
    #   jupyter-client
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.42
    # via -r requirements.in
stack-data==0.6.3
    # via ipython
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
traitlets==5.14.3
    # via
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
typing-extensions==4.14.1
    # via sqlalchemy
tzdata==2025.2
    # via
    #   faker
    #   pandas
wcwidth==0.2.13
    # via prompt-toolkit
